round: move actor-only messages from events.go to actor.go
RefreshVTXORequest, LeaveVTXORequest, and buildVTXORequestFromRefresh
are actor messages (RoundReceivable), not FSM events. Move them to
actor.go where their handlers live, keeping events.go focused on FSM
event types.
